requirements.txt
rich>=13.0.0
prompt_toolkit>=3.0.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
tree-sitter>=0.21.0
tree-sitter-languages>=1.10.2
 openai>=1.3.0
questionary>=2.0.0
